dating_backend/django/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

django
djangorestframework 
djangorestframework-simplejwt 
django-filter
django-environ 
psycopg2-binary
django-cors-headers
autoflake
pillow
drf-yasg
pgvector
gunicorn
torch==2.2.0+cpu
sentence-transformers

channels 
channels-redis
